edit: Update file structure to reorganize responsibilities within microservice
