+++ .gitignore
@@ -1,1 +1,1 @@
-/recipes/.env
+.env
